updated commiting, queries and more variables for $SESSION

+++ assets/dynamics/functions/shopping-card/delete.php
@@ -1,8 +1,8 @@
 <?php
 
-include_once $_SERVER["DOCUMENT_ROOT"] . '/mysql/_.session.php';
+header('Content-type: application/json');
 
-$pdo->beginTransaction();
+include_once $_SERVER["DOCUMENT_ROOT"] . '/mysql/_.session.php';
 
 if (
     isset($_REQUEST['action'], $_REQUEST['id'])
@@ -11,33 +11,58 @@ if (
     && $loggedIn
 ) {
 
-    $id = htmlspecialchars($_REQUEST['id']);
-    $uid = htmlspecialchars($my->id);
+    $debugArray = [];
+
+    // variablize
+    $id = $_REQUEST['id'];
+    $uid = $my->id;
 
     // CHECK EXISTENCE
     $getProduct = $pdo->prepare("SELECT * FROM shopping_card WHERE uid = ? AND pid = ? AND active = '1'");
     $getProduct->execute([$uid, $id]);
+
     if ($getProduct->rowCount() > 0) {
 
-        // UPDATE SCARD
+        $debugArray["productExists"] = true;
+
+        // start mysql transactions
+        $pdo->beginTransaction();
+
+        // update shopping card
         $update = $pdo->prepare("UPDATE shopping_card SET active = '0' WHERE uid = ? AND pid = ?");
-        $update->execute([$uid, $id]);
+        $try = $shop->tryExecute($update, [$uid, $id], $pdo);
 
-        // DELETE RESERVATION
-        $delete = $pdo->prepare("DELETE FROM products_reserved WHERE uid = ? AND pid = ?");
-        $delete->execute([$uid, $id]);
+        if (is_array($try) && $try) {
 
-        if ($update && $delete) {
+            $debugArray["updatedShoppingCard"] = true;
 
-            $pdo->commit();
-            exit('2');
+            // delete reservation
+            $delete = $pdo->prepare("DELETE FROM products_reserved WHERE uid = ? AND pid = ?");
+            $try = $shop->tryExecute($delete, [$uid, $id], $pdo);
+
+            if (is_array($try) && $try) {
+
+                $debugArray["deletedReservations"] = true;
+
+                // update shopping card amount
+                $shoppingCardAmount = $_SESSION["shoppingCardAmount"] - 1;
+                $_SESSION["shoppingCardAmount"]--;
+
+                // create error output
+                $response = [
+                    "status" => true,
+                    "shoppingCardAmount" => $shoppingCardAmount
+                ];
+
+                // commit and exit
+                $pdo->commit();
+                exit(json_encode($response));
+            }
         } else {
-
-            $pdo->rollback();
-            exit('0');
+            exit("0");
         }
     } else {
-        exit('1'); // product doesn't exist
+        exit("1"); // product doesn't exist
     }
 } else {
     exit("0");
